# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Referaa - Refer, mentor, and connect platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # Jobs
...
  /referrals:
    get:
      operationId: listReferrals
      tags: [referrals]
      summary: List referrals for current user (as requester or referrer)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [requester, referrer]
      responses:
        "200":
          description: List of referrals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Referral"
    post:
      operationId: createReferral
      tags: [referrals]
      summary: Request a referral for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralInput"
      responses:
        "201":
          description: Referral request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
  /referrals/{referralId}:
    get:
      operationId: getReferral
      tags: [referrals]
      summary: Get a referral by ID
      parameters:
        - name: referralId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Referral detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
    patch:
      operationId: updateReferral
      tags: [referrals]
      summary: Update referral status (accept/reject/update interview status)
      parameters:
        - name: referralId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralUpdate"
      responses:
        "200":
          description: Updated referral
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"

  # Messages
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        resumeScore:
          type: ["integer", "null"]
        successRate:
          type: ["integer", "null"]
        totalPoints:
          type: integer
        isWorkingProfessional:
          type: boolean
        isConsultant:
          type: boolean
        memberType:
          type: string
          enum: [student, alumni]
        workExperiences:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              role:
                type: ["string", "null"]
              fromYear:
                type: ["string", "null"]
              toYear:
                type: ["string", "null"]
              description:
                type: ["string", "null"]
            required:
              - company
        projects:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: ["string", "null"]
              technologies:
                type: ["string", "null"]
            required:
              - name
        education:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
                enum: [UG, PG, PhD]
              institution:
                type: string
              stream:
                type: ["string", "null"]
              batchYear:
                type: ["integer", "null"]
            required:
              - level
              - institution
        researchPapers:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              publication:
                type: ["string", "null"]
              year:
                type: ["integer", "null"]
              link:
                type: ["string", "null"]
            required:
              - title
        certifications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              issuer:
                type: ["string", "null"]
              year:
                type: ["integer", "null"]
            required:
              - name
        isProfileComplete:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - fullName
        - email
        - skills
        - totalPoints
        - isWorkingProfessional
        - isConsultant
        - memberType
        - workExperiences
        - projects
        - education
        - researchPapers
        - certifications
        - isProfileComplete
        - createdAt
    UserProfileInput:
      type: object
      properties:
        fullName:
          type: string
        mobile:
          type: string
        headline:
          type: string
        bio:
          type: string
        company:
          type: string
        currentRole:
          type: string
        experienceYears:
          type: integer
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
        avatarUrl:
          type: string
        isWorkingProfessional:
          type: boolean
        isConsultant:
          type: boolean
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogoUrl:
          type: ["string", "null"]
        location:
          type: string
        isRemote:
          type: boolean
        isHybrid:
          type: boolean
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryDisclosed:
          type: boolean
        rewardPoints:
          type: integer
        status:
          type: string
          enum: [active, closed, paused]
        posterId:
          type: integer
        poster:
          $ref: "#/components/schemas/UserProfile"
        isSaved:
          type: boolean
        referralCount:
          type: integer
        isOwnJob:
          type: boolean
        myReferralStatus:
          type: ["string", "null"]
          enum: [pending, accepted, rejected, referred, interviewing, hired, rejected_after_interview, null]
        myReferral:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            status:
              type: string
            note:
              type: ["string", "null"]
            totalPointsCredited:
              type: integer
            totalPointsDeducted:
              type: integer
            createdAt:
              type: string
            updatedAt:
              type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - company
        - location
        - isRemote
        - isHybrid
        - description
        - skills
        - salaryDisclosed
        - rewardPoints
        - status
        - posterId
        - poster
        - isSaved
        - referralCount
        - isOwnJob
        - createdAt
...
    Referral:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        job:
          $ref: "#/components/schemas/Job"
        requesterId:
          type: integer
        requester:
          $ref: "#/components/schemas/UserProfile"
        referrerId:
          type: integer
        referrer:
          $ref: "#/components/schemas/UserProfile"
        status:
          type: string
          enum: [pending, accepted, rejected, referred, interviewing, hired, rejected_after_interview]
        interviewStatus:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        rewardTransferred:
          type: boolean
        rewardStagesApplied:
          type: array
          items:
            type: string
        totalPointsDeducted:
          type: integer
        totalPointsCredited:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - jobId
        - job
        - requesterId
        - requester
        - referrerId
        - referrer
        - status
        - rewardTransferred
        - rewardStagesApplied
        - totalPointsDeducted
        - totalPointsCredited
        - createdAt
        - updatedAt
...
    ReferralInput:
      type: object
      properties:
        jobId:
          type: integer
        note:
          type: string
      required:
        - jobId
    ReferralUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, referred, interviewing, hired, rejected_after_interview]
        interviewStatus:
          type: string
        note:
          type: string
...